added private model (listedProducts) to export products to thymeleaf. Converted maxPrice and minPrice to formatted String with two decimal places.

+++ src/main/java/com/brokeshirts/ecom/controllers/DataController.java
@@ -71,7 +71,7 @@ public class DataController {
 
     // ADD INVENTORY ITEM
     @RequestMapping(value="add/item", method = RequestMethod.POST)
-    public String addItem(@RequestParam int productId, @RequestParam int colorId, @RequestParam(value="sizeIds", required = true) int[] sizeIds, @RequestParam("file") MultipartFile file, @RequestParam String sku, @RequestParam double price) {
+    public String addItem(@RequestParam int productId, @RequestParam int colorId, @RequestParam(value="sizeIds", required = true) int[] sizeIds, @RequestParam("file") MultipartFile file, @RequestParam String sku, @RequestParam float price) {
 
         System.out.println("productId: " + productId);
         System.out.println("colorId: " + colorId);
